refactor: update company search logic
