+++ src/main/java/org/scoula/backend/member/controller/CompanyController.java
@@ -1,8 +1,9 @@
 package org.scoula.backend.member.controller;
 
+import java.util.Collections;
 import java.util.List;
 
-import org.scoula.backend.member.dto.CompanySearchResponseDto;
+import org.scoula.backend.member.controller.response.CompanySearchResponseDto;
 import org.scoula.backend.member.service.CompanyService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -26,11 +27,17 @@ public class CompanyController {
 	 * 검색어를 기반으로 회사를 조회하는 메서드
 	 *
 	 * @param query 검색어 (회사명 또는 코드)
-	 * @return 검색된 회사 리스트
+	 * @return 검색된 회사 리스트 또는 400 Bad Request (빈 쿼리인 경우)
 	 */
+
 	@GetMapping("/search")
 	public List<CompanySearchResponseDto> searchCompanies(@RequestParam final String query) {
-		List<CompanySearchResponseDto> companies = companyService.searchCompanies(query);
-		return companies;
+		if (query == null || query.trim().isEmpty()) {
+			log.info("빈 검색어가 입력되었습니다.");
+			return Collections.emptyList();
+		}
+		return companyService.searchCompanies(query);
 	}
+
 }
+
