+++ src/main/java/org/scoula/backend/order/controller/response/OrderSummaryResponse.java
@@ -10,4 +10,5 @@ public class OrderSummaryResponse {
 	private final String companyCode;
 	private final Integer sellCount;
 	private final Integer buyCount;
+
 }
